+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## [2.6.0] - 2024-01-26
+
+* Adding widget to DropdownField child.
+
 ## [2.5.1] - 2024-01-22
 
 * Rolling back ModelUtils.
-* "Renaming uiBuilder to builder.
+* Renaming uiBuilder to builder.
 * Adding builder and consumer to actions and rowActions attributes in
   AbstractList.
 * Creating ModelUtils methods: